fix: 修复 Debug x64 编译 45 个 warning（正确修复，无 pragma 压制）
整理为 5 类修复：
1. 删除未使用变量（Client.cpp, Server.cpp, RecieveSocket.cpp）
2. 类型转换 static_cast/reinterpret_cast（Clip.cpp, DeleteClipData.cpp, ImageViewer.cpp,
   HyperLink.cpp, QPasteWnd.cpp, QPasteWndThread.cpp, ToolTipEx.cpp, SendKeys.cpp,
   SymbolEdit.cpp, Options.cpp, Misc.cpp, RecieveSocket.cpp, Client.cpp, ChaiScriptOnCopy.cpp）
3. 弃用 API 替换（Misc.cpp 的 GetVersionEx → 简化逻辑; SendKeys.cpp 的 GetVersion → 移除;
   Options.cpp 的 wcstok → _tcstok_s）
4. 宏冲突（UnicodeMacros.h 删除 STRTOK; Misc.cpp OUT/IN → enum）
5. 类型库导入冲突（StdAfx.h 加 rename FindText→FindTextW）

Md5.cpp 删除废弃 register 关键字。
EditWithButton.cpp 因 MFC 内部实现无法从外部修复，仅此一处用 #pragma 局部压制。
未引入新括号不平衡（预存的 Clip.cpp+1, ToolTipEx.cpp+2, SendKeys.cpp+1 未触碰）。

+++ src/Clip.cpp
@@ -2113,12 +2113,12 @@ BOOL CClip::SaveFormats(CString *unicode, CStringA *asci, CStringA *rtf, BOOL up
 
 	if (cf_dibBytes != nullptr && cf_dibBytes->size() > 0)
 	{
-		AddFormat(CF_DIB, cf_dibBytes->data(), cf_dibBytes->size(), false);
+		AddFormat(CF_DIB, cf_dibBytes->data(), static_cast<UINT>(cf_dibBytes->size()), false);
 	}
 
 	if (pngBytes != nullptr && pngBytes->size() > 0)
 	{
-		AddFormat(theApp.m_PNG_Format, pngBytes->data(), pngBytes->size(), false);
+		AddFormat(theApp.m_PNG_Format, pngBytes->data(), static_cast<UINT>(pngBytes->size()), false);
 	}
 
 	if (rtf != nullptr)
